add logic to switch between login and register form
